LibGL: Introduce a class based gl context approach
This introduces a base GLContext interface and a software based
renderer implementation. The global glXXX functions serve as a
c-style interface to the context instance.
