+++ Meta/CMake/all_the_debug_macros.cmake
@@ -173,6 +173,7 @@ set(SYSCALL_1_DEBUG ON)
 set(RSA_PARSE_DEBUG ON)
 set(LINE_EDITOR_DEBUG ON)
 set(LANGUAGE_SERVER_DEBUG ON)
+set(GL_DEBUG ON)
 
 # False positive: DEBUG is a flag but it works differently.
 # set(DEBUG ON)
