+++ Userland/Libraries/LibGL/GLContext.h
@@ -1,48 +1,33 @@
 /*
  * Copyright (c) 2021, Jesse Buhagiar <jooster669@gmail.com>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ * Copyright (c) 2021, Stephan Unverwerth <s.unverwerth@gmx.de>
+ * 
+ * SPDX-License-Identifier: BSD-2-Clause
  */
+
 #pragma once
 
 #include "GL/gl.h"
-#include "GLStruct.h"
 
-#include <AK/Vector.h>
-#include <LibGfx/Matrix4x4.h>
+class GLContext {
+public:
+    virtual ~GLContext() { }
 
-struct GLContext {
-    GLenum curr_draw_mode;
-    GLenum curr_matrix_mode;
-    FloatMatrix4x4 projection_matrix;
-    FloatMatrix4x4 model_view_matrix;
-
-    Vector<FloatMatrix4x4> projection_matrix_stack;
-    Vector<FloatMatrix4x4> model_view_matrix_stack;
-
-    color_t clear_color = { 0.0f, 0.0f, 0.0f, 0.0f };
-    color_t curr_vertex_color = { 255.0f, 255.0f, 255.0f, 255.0f };
+    virtual void gl_begin(GLenum mode) = 0;
+    virtual void gl_clear(GLbitfield mask) = 0;
+    virtual void gl_clear_color(GLclampf red, GLclampf green, GLclampf blue, GLclampf alpha) = 0;
+    virtual void gl_color(GLdouble r, GLdouble g, GLdouble b, GLdouble a) = 0;
+    virtual void gl_end() = 0;
+    virtual void gl_frustum(GLdouble left, GLdouble right, GLdouble bottom, GLdouble top, GLdouble nearVal, GLdouble farVal) = 0;
+    virtual GLubyte* gl_get_string(GLenum name) = 0;
+    virtual void gl_load_identity() = 0;
+    virtual void gl_matrix_mode(GLenum mode) = 0;
+    virtual void gl_push_matrix() = 0;
+    virtual void gl_pop_matrix() = 0;
+    virtual void gl_rotate(GLdouble angle, GLdouble x, GLdouble y, GLdouble z) = 0;
+    virtual void gl_translate(GLdouble x, GLdouble y, GLdouble z) = 0;
+    virtual void gl_vertex(GLdouble x, GLdouble y, GLdouble z, GLdouble w) = 0;
+    virtual void gl_viewport(GLint x, GLint y, GLsizei width, GLsizei height) = 0;
 };
 
-extern GLContext* g_gl_state;
-extern FloatMatrix4x4 g_curr_matrix;
+extern GLContext* g_gl_context;
